+++ core/src/scheduler/vm.rs
@@ -21,6 +21,7 @@ use alloc::{
     vec::Vec,
 };
 use core::{cell::RefCell, convert::TryInto, fmt};
+use redshirt_syscalls_interface::{Pid, MessageId};
 use smallvec::SmallVec;
 
 /// WASMI state machine dedicated to a process.
