+++ interfaces/syscalls/src/lib.rs
@@ -94,6 +94,8 @@ pub use interface_message::{emit_answer, next_interface_message, InterfaceMessag
 pub use response::{message_response, message_response_sync_raw, MessageResponseFuture};
 pub use traits::{Decode, Encode, EncodedMessage};
 
+use core::fmt;
+
 mod block_on;
 mod emit;
 mod interface_message;
@@ -101,3 +103,49 @@ mod response;
 mod traits;
 
 pub mod ffi;
+
+/// Identifier of a running process within a core.
+// TODO: move to a Pid module?
+#[derive(Copy, Clone, PartialEq, Eq, Hash)]
+pub struct Pid(u64);
+
+impl From<u64> for Pid {
+    fn from(id: u64) -> Pid {
+        Pid(id)
+    }
+}
+
+impl From<Pid> for u64 {
+    fn from(pid: Pid) -> u64 {
+        pid.0
+    }
+}
+
+impl fmt::Debug for Pid {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "#{}", self.0)
+    }
+}
+
+/// Identifier of a message to answer.
+// TODO: move to a MessageId module?
+#[derive(Copy, Clone, PartialEq, Eq, Hash)]
+pub struct MessageId(u64);
+
+impl From<u64> for MessageId {
+    fn from(id: u64) -> MessageId {
+        MessageId(id)
+    }
+}
+
+impl From<MessageId> for u64 {
+    fn from(pid: MessageId) -> u64 {
+        pid.0
+    }
+}
+
+impl fmt::Debug for MessageId {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "#{}", self.0)
+    }
+}
