+++ kernel/hosted-wasi/Cargo.toml
@@ -12,5 +12,6 @@ hashbrown = { version = "0.6.0", default-features = false }
 redshirt-core = { path = "../../core", default-features = false }
 redshirt-random-interface = { path = "../../interfaces/random", default-features = false }
 redshirt-stdout-interface = { path = "../../interfaces/stdout", default-features = false }
+redshirt-syscalls-interface = { path = "../../interfaces/syscalls", default-features = false }
 redshirt-time-interface = { path = "../../interfaces/time", default-features = false }
 parity-scale-codec = { version = "1.1.0", default-features = false }
