+++ kernel/standalone/Cargo.toml
@@ -18,6 +18,7 @@ parity-scale-codec = { version = "1.0.5", default-features = false }
 redshirt-core = { path = "../../core" }
 redshirt-hardware-interface = { path = "../../interfaces/hardware", default-features = false }
 redshirt-stdout-interface = { path = "../../interfaces/stdout", default-features = false }
+redshirt-syscalls-interface = { path = "../../interfaces/syscalls", default-features = false }
 redshirt-wasi-hosted = { path = "../hosted-wasi" }
 spin = "0.5.2"
